requirements.txt
web3==5.31.1
py-solc-x==1.1.0
python-dotenv==0.21.0